Add new tempfile model and improve core models.
I have added a new tempfile model to store and look after uploaded files while we wait for the user to provide us with more information.

In addtion the core models have been in proved. In perticular we now store the md5sum of files so that we know when the same file has been uploaded twice.