Improve upload views to improve file handleing.
The upload views have had a huge improvement. File uploading is handled in one view, with files being temp stored on the filesystem with db referances (so that we can track the file using sessions).

The file details is now completely seperate to the file upload, there just needs to be a file referanced in the  current session.