Add the new PDF creation task.
I have now written the celery task for creating PDF's from uploaded files, and it WORKS. It has been written in a manner which does not require django to run. This means the task could be run on a stand-allown celery instance without the rest of the site tagging along.

Currently it requires that the conversion service is running, also urls are hard coded at the moment.