Add call to new pdf convert task, and change metadata use and add logging.
We now use python logging to handle logging.

We now queue a pdf convert task after the document details have been confirmed by the user.

The title sent back to the user in the form is now the file name, not the metadata title, this is because most of the files being used dont have a proper title in meta-data.