Add has_pdf method and have django serve derived pdf files.
By appending a type to the document url (e.g. document/1/pdf/) it is now possiable to get a derived file type (based on appended string). If django cant find the correct type it will serve the original.

I addition django will serve the original if the original is a pdf and requested type is a pdf.

The document model now has a has_pdf method which returns true if it is possiable for django to serve a pdf.

The search template has been updated to take advantage of both of these changes by providing links to the pdf types if possiable.