Big changes. New view for serving dynamicly scaled images, and setting to make it work. New tag model. New faced search front end that uses the new tag model. Some helper functions to make using the models easier.
The new view serves "thumbnail" images that are scaled on the fly, and cached to disk. Has settings, where to store the cache and what image widths are allowed.

The new tag model is quite simple really, just a title, description and many-to-many field.

The helper functions for the models are a simple set, one to generate md5's from a models assciated file, and a whole load of __unicode__ functions.

The search front end now uses facets, and is faceted on tags at the moment.